deleted StoryScreen.tsx and moved file changes to story.tsx
